split minio reader and writer
