+++ source/minio/minio.go
@@ -2,185 +2,15 @@ package minio
 
 import (
 	"context"
-	"errors"
-	"io"
 
 	"github.com/minio/minio-go/v7"
-
-	"github.com/hangxie/parquet-go/v2/source"
-)
-
-// Compile time check that *minioFile implement the source.ParquetFileReader and source.ParquetFileWriter interface.
-var (
-	_ source.ParquetFileReader = (*minioFile)(nil)
-	_ source.ParquetFileWriter = (*minioFile)(nil)
 )
 
 // minioFile is ParquetFile for MinIO S3 API
 type minioFile struct {
-	ctx    context.Context
-	client *minio.Client
-	offset int64
-	whence int
-
-	// write-related fields
-	pipeReader *io.PipeReader
-	pipeWriter *io.PipeWriter
-
-	// read-related fields
-	fileSize   int64
-	downloader *minio.Object
-
-	err        error
+	ctx        context.Context
+	client     *minio.Client
 	bucketName string
 	key        string
-}
-
-var (
-	errWhence        = errors.New("Seek: invalid whence")
-	errInvalidOffset = errors.New("Seek: invalid offset")
-)
-
-// NewS3FileWriterWithClient is the same as NewMinioFileWriter but allows passing
-// your own S3 client.
-func NewS3FileWriterWithClient(
-	ctx context.Context,
-	s3Client *minio.Client,
-	bucket string,
-	key string,
-) (source.ParquetFileWriter, error) {
-	file := &minioFile{
-		ctx:        ctx,
-		client:     s3Client,
-		bucketName: bucket,
-		key:        key,
-	}
-
-	return file.Create(key)
-}
-
-// NewS3FileReaderWithClient is the same as NewMinioFileReader but allows passing
-// your own S3 client
-func NewS3FileReaderWithClient(ctx context.Context, s3Client *minio.Client, bucket, key string) (source.ParquetFileReader, error) {
-	file := &minioFile{
-		ctx:        ctx,
-		client:     s3Client,
-		bucketName: bucket,
-		key:        key,
-	}
-
-	return file.Open(key)
-}
-
-// Seek tracks the offset for the next Read. Has no effect on Write.
-func (s *minioFile) Seek(offset int64, whence int) (int64, error) {
-	if whence < io.SeekStart || whence > io.SeekEnd {
-		return 0, errWhence
-	}
-
-	if s.fileSize > 0 {
-		switch whence {
-		case io.SeekStart:
-			if offset < 0 || offset > s.fileSize {
-				return 0, errInvalidOffset
-			}
-		case io.SeekCurrent:
-			offset += s.offset
-			if offset < 0 || offset > s.fileSize {
-				return 0, errInvalidOffset
-			}
-		case io.SeekEnd:
-			if offset > -1 || -offset > s.fileSize {
-				return 0, errInvalidOffset
-			}
-		}
-	}
-
-	s.offset = offset
-	s.whence = whence
-	return s.offset, nil
-}
-
-// Read up to len(p) bytes into p and return the number of bytes read
-func (s *minioFile) Read(p []byte) (n int, err error) {
-	if s.fileSize > 0 && s.offset >= s.fileSize {
-		return 0, io.EOF
-	}
-
-	bytesDownloaded, err := s.downloader.ReadAt(p, s.offset)
-	if err != nil {
-		return 0, err
-	}
-
-	s.offset += int64(bytesDownloaded)
-	return bytesDownloaded, err
-}
-
-// Write len(p) bytes from p to the Minio data stream
-func (s *minioFile) Write(p []byte) (n int, err error) {
-	// prevent further writes upon error
-	bytesWritten, writeError := s.pipeWriter.Write(p)
-	if writeError != nil {
-		s.err = writeError
-		s.pipeWriter.CloseWithError(err)
-		return 0, writeError
-	}
-
-	return bytesWritten, nil
-}
-
-// Close signals write completion and cleans up any
-// open streams. Will block until pending uploads are complete.
-func (s *minioFile) Close() error {
-	var err error
-
-	if s.pipeWriter != nil {
-		if err = s.pipeWriter.Close(); err != nil {
-			return err
-		}
-	}
-
-	return err
-}
-
-// Open creates a new Minio File instance to perform concurrent reads
-func (s *minioFile) Open(name string) (source.ParquetFileReader, error) {
-	// new instance
-	pf := &minioFile{
-		ctx:        s.ctx,
-		client:     s.client,
-		bucketName: s.bucketName,
-		key:        name,
-		offset:     0,
-	}
-	// init object info
-	downloader, err := s.client.GetObject(s.ctx, s.bucketName, s.key, minio.GetObjectOptions{})
-	if err != nil {
-		return pf, err
-	}
-	info, err := downloader.Stat()
-	if err != nil {
-		return pf, err
-	}
-	s.downloader = downloader
-	s.fileSize = info.Size
-	return pf, nil
-}
-
-// Create creates a new Minio File instance to perform writes
-func (s *minioFile) Create(key string) (source.ParquetFileWriter, error) {
-	pf := &minioFile{
-		ctx:        s.ctx,
-		client:     s.client,
-		bucketName: s.bucketName,
-		key:        key,
-	}
-	pr, pw := io.Pipe()
-	_, err := s.client.PutObject(s.ctx, s.bucketName, s.key, pr, -1, minio.PutObjectOptions{})
-	if err != nil {
-		return pf, err
-	}
-	pf.pipeReader = pr
-	pf.pipeWriter = pw
-	return pf, nil
+	err        error
 }
